# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "DFD API",
        "contact": {
            "name": "Seokho Song",
            "url": "http://github.com/devsdkGIN_MODE"
        },
        "license": {
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "devsdk.net/api/dfd",
    "basePath": "/",
    "paths": {
...
        "/v1/announce": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "write announce\nPermission : **announce.post**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/announce"
                ],
                "summary": "Write announce",
                "parameters": [
                    {
                        "description": "username",
                        "name": "bodymap",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.RequestBodyAnnouncePost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/announce/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get specific announce by announce id\nPermission : **announce.get**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/announce"
                ],
                "summary": "Get announce",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Announce ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "announce": {
                                            "$ref": "#/definitions/models.Announce"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete announce by announce id\nPermission : **announce.delete**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/announce"
                ],
                "summary": "Delete announce",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Announce ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Edit spcific announce written by me\nPermission : **announce.get**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/announce"
                ],
                "summary": "Patch Announce",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Announce ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.RequestBodyAnnouncePost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/announces/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all announce list sorted by creation time\nPermission : **announces.get**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/announce"
                ],
                "summary": "Get all announce list",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Announce"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/announces/current": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get announce list greater then target-date from NOW sorted by creation time\nPermission : **announces.get**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/announce"
                ],
                "summary": "Get current announce list",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Announce"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/announces/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get announce list written by me\nPermission : **announces.get**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api/v1/announce"
                ],
                "summary": "Get own announce list",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docmodels.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Announce"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseBadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Request. If token is expired, **token_expired** filed must be set true",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseUnauthorized"
                        }
                    },
                    "403": {
                        "description": "You don't have permission",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "404": {
                        "description": "Cannt found user",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docmodels.ResponseInternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docmodels.RequestBodyAnnouncePost": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This field is optional"
                },
                "target_date": {
                    "type": "string",
                    "format": "date-time"
                },
                "title": {
                    "type": "string",
                    "example": "Title"
                }
            }
        },
...
        "docmodels.ResponseBadRequest": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Bad request message"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
...
        "docmodels.ResponseInternalServerError": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
...
        "docmodels.ResponseNotFound": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Not found message"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "docmodels.ResponseSuccess": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "docmodels.ResponseUnauthorized": {
            "type": "object",
            "properties": {
                "messsage": {
                    "type": "string",
                    "example": "Request Unauthorized message"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "token_expired": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.Announce": {
            "type": "object",
            "properties": {
                "authorId": {
                    "type": "string",
                    "example": "5c06bc289b9bdd2381263063"
                },
                "created": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "this project is awesome!"
                },
                "id": {
                    "type": "string",
                    "example": "6006bc289b9bdd2381263063"
                },
                "modified": {
                    "type": "string"
                },
                "targetDate": {
                    "type": "string",
                    "example": "2021-01-19T11:01:19+00:00"
                },
                "title": {
                    "type": "string",
                    "example": "awesome title"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}